+++ sdk/src/main/java/com/mapbox/maps/StyleObserver.kt
@@ -20,9 +20,13 @@ internal class StyleObserver(
   private val pixelRatio: Float
 ) : OnStyleLoadedListener, OnMapLoadErrorListener, OnStyleDataLoadedListener {
 
-  private var loadStyleListener: Style.OnStyleLoaded? = null
+  private var userStyleLoadedListener: Style.OnStyleLoaded? = null
+
+  private var styleDataStyleLoadedListener: Style.OnStyleLoaded? = null
+  private var styleDataSpritesLoadedListener: Style.OnStyleLoaded? = null
+  private var styleDataSourcesLoadedListener: Style.OnStyleLoaded? = null
+
   private var loadStyleErrorListener: OnMapLoadErrorListener? = null
-  private var loadStyleTransitionOptions: TransitionOptions? = null
 
   private val getStyleListeners = CopyOnWriteArraySet<Style.OnStyleLoaded>()
   private var loadedStyle: Style? = null
@@ -39,13 +43,17 @@ internal class StyleObserver(
    * NOTE : listener is invoked only once after successful style load.
    */
   fun setLoadStyleListener(
-    transitionOptions: TransitionOptions?,
-    loadedListener: Style.OnStyleLoaded?,
+    userOnStyleLoaded: Style.OnStyleLoaded?,
+    styleDataStyleLoadedListener: Style.OnStyleLoaded?,
+    styleDataSpritesLoadedListener: Style.OnStyleLoaded?,
+    styleDataSourcesLoadedListener: Style.OnStyleLoaded?,
     onMapLoadErrorListener: OnMapLoadErrorListener?
   ) {
-    loadStyleTransitionOptions = transitionOptions
-    loadStyleListener = loadedListener
-    loadStyleErrorListener = onMapLoadErrorListener
+    this.userStyleLoadedListener = userOnStyleLoaded
+    this.styleDataStyleLoadedListener = styleDataStyleLoadedListener
+    this.styleDataSpritesLoadedListener = styleDataSpritesLoadedListener
+    this.styleDataSourcesLoadedListener = styleDataSourcesLoadedListener
+    this.loadStyleErrorListener = onMapLoadErrorListener
   }
 
   /**
@@ -60,13 +68,11 @@ internal class StyleObserver(
    * Invoked when a style has loaded
    */
   override fun onStyleLoaded(eventData: StyleLoadedEventData) {
-    val style = Style(nativeMap, pixelRatio)
-    loadedStyle?.markInvalid()
-    loadedStyle = style
+    val style = loadedStyle ?: throw MapboxMapException("Style is not initialized on onStyleLoaded callback!")
     styleLoadedListener.onStyleLoaded(style)
 
-    loadStyleListener?.onStyleLoaded(style)
-    loadStyleListener = null
+    userStyleLoadedListener?.onStyleLoaded(style)
+    userStyleLoadedListener = null
 
     getStyleListeners.forEach { listener ->
       listener.onStyleLoaded(style)
@@ -83,21 +89,35 @@ internal class StyleObserver(
   }
 
   override fun onStyleDataLoaded(eventData: StyleDataLoadedEventData) {
-    // style data arrives in following order: STYLE, SOURCES, SPRITE
-    // transition options must be applied after style but before sprite and sources to take effect
-    loadStyleTransitionOptions?.let {
-      if (eventData.type == StyleDataType.STYLE) {
-        nativeMap.styleTransition = it
-        // per gl-native docs style transition options should be reset for a new style so resetting them here
-        loadStyleTransitionOptions = null
+    when (eventData.type) {
+      StyleDataType.STYLE -> {
+        loadedStyle?.markInvalid()
+        loadedStyle = Style(nativeMap, pixelRatio).also {
+          styleDataStyleLoadedListener?.onStyleLoaded(it)
+        }
+        styleDataStyleLoadedListener = null
+      }
+      StyleDataType.SPRITE -> {
+        styleDataSpritesLoadedListener?.onStyleLoaded(
+          loadedStyle ?: throw MapboxMapException("Style is not initialized yet although SPRITES event has arrived!")
+        )
+        styleDataSpritesLoadedListener = null
+      }
+      StyleDataType.SOURCES -> {
+        styleDataSourcesLoadedListener?.onStyleLoaded(
+          loadedStyle ?: throw MapboxMapException("Style is not initialized yet although SOURCES event has arrived!")
+        )
+        styleDataSourcesLoadedListener = null
       }
     }
   }
 
   fun onDestroy() {
-    loadStyleListener = null
+    userStyleLoadedListener = null
+    styleDataStyleLoadedListener = null
+    styleDataSpritesLoadedListener = null
+    styleDataSourcesLoadedListener = null
     loadStyleErrorListener = null
-    loadStyleTransitionOptions = null
     loadedStyle?.markInvalid()
     loadedStyle = null
     getStyleListeners.clear()